Upload traj for QR without load
The uploaded trajectory is for a very small QR since I did not update the object properties for the QR wihtout load in Quadrotor.m. If we change the object properties, some changes need to be made in the traj_gen_QR file to automatically read them from the obj.  I thought this has not the highest priority since we only need the QRL model in the end.